refactor: split codegen runtime metadata
